# docs/03-architecture/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareFlow API",
    "version": "1.0.0",
    "description": "Machine-readable contract for the CareFlow clinic operations API v1. All paths are mounted under /api/v1 and use the shared success, list and error envelopes."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local API server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in with email and password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/AuthSession" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Đăng ký tài khoản bệnh nhân",
        "security": [],
        "requestBody": { "$ref": "#/components/requestBodies/PatientRegistration" },
        "responses": {
          "201": { "$ref": "#/components/responses/AuthSession" },
          "400": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Đổi mật khẩu của tài khoản hiện tại",
        "requestBody": { "$ref": "#/components/requestBodies/ChangePassword" },
        "responses": {
          "201": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log out the current bearer session",
        "responses": {
          "201": { "$ref": "#/components/responses/Success" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get the current authenticated user",
        "responses": {
          "200": { "$ref": "#/components/responses/CurrentUser" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "requestBodies": {
...
      "PatientRegistration": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/PatientRegistrationRequest" }
          }
        }
      },
      "ChangePassword": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" }
          }
        }
      },
...
    },
    "responses": {
      "Success": {
        "description": "Success envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/SuccessEnvelope" }
          }
        }
      },
...
      "Error": {
        "description": "Error envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "AuthSession": {
        "description": "Authenticated user and bearer session",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/AuthSessionEnvelope" }
          }
        }
      },
      "CurrentUser": {
        "description": "Current authenticated user",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/CurrentUserEnvelope" }
          }
        }
      },
...
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": ["data", "meta"],
        "properties": {
          "data": { "type": "object", "additionalProperties": true },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": ["error", "meta"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "$ref": "#/components/schemas/ErrorCode" },
              "message": { "type": "string" },
              "fields": {
                "type": "object",
                "additionalProperties": { "type": "string" }
              }
            }
          },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        }
      },
      "ResponseMeta": {
        "type": "object",
        "required": ["requestId"],
        "properties": { "requestId": { "type": "string" } }
      },
...
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "PatientRegistrationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["displayName", "email", "phone", "password"],
        "properties": {
          "displayName": { "type": "string", "minLength": 1, "maxLength": 200 },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string", "minLength": 1, "maxLength": 30 },
          "password": { "type": "string", "minLength": 8, "maxLength": 72, "description": "Maximum 72 UTF-8 bytes." }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["currentPassword", "newPassword"],
        "properties": {
          "currentPassword": { "type": "string", "minLength": 1, "maxLength": 72, "description": "Maximum 72 UTF-8 bytes." },
          "newPassword": { "type": "string", "minLength": 8, "maxLength": 72, "description": "Maximum 72 UTF-8 bytes." }
        }
      },
      "AuthSessionEnvelope": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": ["sessionToken", "currentUser", "linkedProfile"],
                "properties": {
                  "sessionToken": { "type": "string" },
                  "currentUser": { "$ref": "#/components/schemas/CurrentUser" },
                  "linkedProfile": {
                    "$ref": "#/components/schemas/LinkedProfile"
                  }
                }
              }
            }
          }
        ]
      },
      "CurrentUserEnvelope": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": ["currentUser", "linkedProfile"],
                "properties": {
                  "currentUser": { "$ref": "#/components/schemas/CurrentUser" },
                  "linkedProfile": {
                    "$ref": "#/components/schemas/LinkedProfile"
                  }
                }
              }
            }
          }
        ]
      },
...
      "CurrentUser": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "displayName", "email", "role", "status"],
        "properties": {
          "id": { "type": "string" },
          "displayName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "$ref": "#/components/schemas/UserRole" },
          "status": { "enum": ["active", "inactive", "locked"] }
        }
      },
...
      "LinkedProfile": {
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": false,
            "required": ["type", "id"],
            "properties": {
              "type": { "enum": ["patient", "doctor", "staff"] },
              "id": { "type": "string" }
            }
          },
          { "type": "null" }
        ]
      },
...
      "UserRole": {
        "enum": ["patient", "doctor", "receptionist", "nurse", "admin"]
      },
...
      "ErrorCode": {
        "enum": [
          "UNAUTHENTICATED",
          "FORBIDDEN",
          "VALIDATION_ERROR",
          "NOT_FOUND",
          "APPOINTMENT_CONFLICT",
          "INVALID_STATUS_TRANSITION",
          "RESOURCE_IN_USE",
          "RATE_LIMITED",
          "INTERNAL_ERROR"
        ]
      }
    }
  }
}